Seperate scraper code from index cli file
