+++ src/dev-cli/config.ts
@@ -6,7 +6,11 @@ export function getConfig() {
     throw new Error('Missing MOVIE_WEB_TMDB_API_KEY environment variable');
   }
 
+  let proxyUrl: undefined | string = process.env.MOVIE_WEB_PROXY_URL;
+  proxyUrl = !proxyUrl ? undefined : proxyUrl;
+
   return {
     tmdbApiKey,
+    proxyUrl,
   };
 }
